Created instance_registry and os crates (#730)
