+++ mosaicod/Cargo.toml
@@ -25,6 +25,8 @@ mosaicod-rw = { path = "crates/mosaicod-rw" }
 mosaicod-grpc = { path = "crates/mosaicod-grpc" }
 mosaicod-grpc-flight = { path = "crates/mosaicod-grpc-flight" }
 mosaicod-grpc-common = { path = "crates/mosaicod-grpc-common" }
+mosaicod-instance-registry = { path = "crates/mosaicod-instance-registry" }
+mosaicod-os = { path = "crates/mosaicod-os" }
 mosaicod-store = { path = "crates/mosaicod-store" }
 mosaicod-task = { path = "crates/mosaicod-task" }
