+++ mosaicod/crates/mosaicod-instance-registry/src/lib.rs
@@ -1,0 +1,2 @@
+mod heartbeat;
+pub use heartbeat::*;
