+++ mosaicod/crates/mosaicod-task/Cargo.toml
@@ -12,6 +12,8 @@ name = "mosaicod_task"
 [dependencies]
 mosaicod-core = { workspace = true }
 mosaicod-db = { workspace = true, features = ["postgres"] }
+mosaicod-os = { workspace = true }
+mosaicod-instance-registry = { workspace = true }
 mosaicod-store = { workspace = true }
 mosaicod-rw = { workspace = true }
 mosaicod-facade = { workspace = true }
@@ -21,7 +23,6 @@ tokio = { workspace = true }
 tokio-util = { workspace = true }
 datafusion = { workspace = true }
 futures = { workspace = true }
-hostname = { workspace = true }
 
 [dev-dependencies]
 mosaicod-db = { workspace = true, features = ["postgres", "testing"] }
