+++ mosaicod/crates/mosaicod-task/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod cleanup;
 pub use cleanup::*;
 
-mod instance;
-
 pub mod store_optimizer;
 pub use store_optimizer::*;
